refactor: extract tab prototype types
