+++ src/factory/index.ts
@@ -1,1 +1,2 @@
 export * from './TabFactory';
+export * from './TabPrototype';
